Creating an images folder in the main
Whilst working on the update-read-me branch I have encountered an error. Images are not displaying on the readme file, however the path links to the image files are correct. This is potentially a bug due to the images being saved in a folder in the update-read-me branch so I will be adding the images folder to the parent directory.